refactor: split latex_header.py into latex_escape and latex_header modules (#242)
Extract LaTeX escaping/validation functions (_escape_latex, _validate_latex_value,
_validate_header_footer_values, _truncate_title, _substitute_placeholders) into
a new latex_escape.py module, keeping block builders in latex_header.py. Both
files are now well under the 300-line limit.
