+++ obsidian_export/pipeline/latex_header.py
@@ -1,16 +1,13 @@
 """Generate rendered LaTeX header from style config and template."""
 
-import re
 from pathlib import Path
 
 from obsidian_export.config import HeadingStyle, StyleConfig, TitleStyle
-from obsidian_export.exceptions import UnsafeLatexError
-
-_DANGEROUS_LATEX_RE = re.compile(
-    r"\\(?:input|include|write\d*|immediate|openin|openout|read|closein|closeout"
-    r"|catcode|def|edef|gdef|xdef|let|csname|newwrite|ShellEscape|directlua"
-    r"|luaexec|luadirect)(?![a-zA-Z])",
-    re.IGNORECASE,
+from obsidian_export.pipeline.latex_escape import (
+    _escape_latex,
+    _substitute_placeholders,
+    _validate_header_footer_values,
+    _validate_latex_value,
 )
 
 
@@ -70,57 +67,6 @@ def render_header(style: StyleConfig, template_path: Path, title: str) -> str:
         title_style_block=title_style_block,
         code_block=code_block,
     )
-
-
-def _truncate_title(title: str) -> str:
-    """Shorten title for header use — cut before first em-dash or colon."""
-    for sep in (" — ", " – ", " - ", ": "):
-        pos = title.find(sep)
-        if pos != -1:
-            return title[:pos].strip()
-    return title
-
-
-def _escape_latex(text: str) -> str:
-    """Escape LaTeX special characters in plain text for safe preamble insertion."""
-    # Order matters: & must come before others that might produce &
-    replacements = [
-        ("\\", "\\textbackslash{}"),
-        ("&", "\\&"),
-        ("%", "\\%"),
-        ("$", "\\$"),
-        ("#", "\\#"),
-        ("_", "\\_"),
-        ("{", "\\{"),
-        ("}", "\\}"),
-        ("~", "\\textasciitilde{}"),
-        ("^", "\\textasciicircum{}"),
-    ]
-    for char, escaped in replacements:
-        text = text.replace(char, escaped)
-    return text
-
-
-def _substitute_placeholders(value: str, title: str, logo_path: str) -> str:
-    """Replace {doc_title} and {logo_path} in a header/footer config string."""
-    if not value:
-        return value
-    short_title = _escape_latex(_truncate_title(title))
-    return value.replace("{doc_title}", short_title).replace("{logo_path}", logo_path)
-
-
-def _validate_latex_value(latex: str, field_name: str) -> None:
-    """Reject latex values containing dangerous macros.
-
-    Raises UnsafeLatexError if the value contains macros that could read files,
-    execute shell commands, or redefine TeX internals.
-    """
-    if _DANGEROUS_LATEX_RE.search(latex):
-        msg = (
-            f"Config field '{field_name}' contains a dangerous LaTeX macro: "
-            f"'{latex}'. Remove or replace it with a safe alternative."
-        )
-        raise UnsafeLatexError(msg)
 
 
 def _build_unicode_char_block(unicode_chars: tuple[tuple[str, str], ...]) -> str:
@@ -198,7 +144,6 @@ def _build_heading_styles_block(heading_styles: tuple[HeadingStyle, ...]) -> str
         if h.color:
             parts.append(f"\\color{{{_escape_latex(h.color)}}}")
         fmt = "".join(parts)
-        # For the content argument (last {}), use \\MakeUppercase if uppercase
         content_arg = "{\\MakeUppercase}" if h.uppercase else "{}"
         lines.append(f"\\titleformat{{\\{h.level}}}\n  {{{fmt}}}\n  {{\\the{h.level}}}{{1em}}{content_arg}")
     return "\n\n".join(lines)
@@ -246,16 +191,6 @@ def _build_code_block(code_fontsize: str) -> str:
         f"\\DefineVerbatimEnvironment{{verbatim}}{{Verbatim}}"
         f"{{breaklines=true, fontsize={cmd}}}"
     )
-
-
-def _validate_header_footer_values(fields: dict[str, str]) -> None:
-    """Validate all header/footer values for dangerous LaTeX macros.
-
-    Raises UnsafeLatexError if any value contains a dangerous macro.
-    """
-    for field_name, value in fields.items():
-        if value:
-            _validate_latex_value(value, field_name)
 
 
 def _build_header_footer_block(
